+++ smac/facade/smac_facade.py
@@ -70,7 +70,7 @@ class SMAC(object):
                  initial_configurations: typing.List[Configuration]=None,
                  stats: Stats=None,
                  restore_incumbent: Configuration=None,
-                 rng: np.random.RandomState=None,
+                 rng: typing.Union[np.random.RandomState, int]=None,
                  smbo_class: SMBO=None,
                  run_id: int=1,
                  random_configuration_chooser=None):
@@ -133,7 +133,28 @@ class SMAC(object):
 
         aggregate_func = average_cost
 
-        self.output_dir = create_output_directory(scenario, run_id)
+        self.scenario = scenario
+        self.output_dir = ""
+        if not restore_incumbent:
+            self.output_dir = create_output_directory(scenario, run_id)
+        elif scenario.output_dir is not None:
+            # output-directory is created in CLI when restoring from a
+            # folder. calling the function again in the facade results in two
+            # folders being created: run_X and run_X.OLD. if we are
+            # restoring, the output-folder exists already and we omit creating it,
+            # but set the self-output_dir to the dir.
+            # necessary because we want to write traj to new output-dir in CLI.
+            self.output_dir = os.path.join(scenario.output_dir,
+                                           "run_%d" % (run_id))
+        if (
+            scenario.deterministic is True
+            and getattr(scenario, 'tuner_timeout', None) is None
+            and scenario.run_obj == 'quality'
+        ):
+            self.logger.info('Optimizing a deterministic scenario for '
+                             'quality without a tuner timeout - will make '
+                             'SMAC deterministic!')
+            scenario.intensification_percentage = 1e-10
         scenario.write()
 
         # initialize stats object
@@ -165,16 +186,16 @@ class SMAC(object):
         # initial EPM
         types, bounds = get_types(scenario.cs, scenario.feature_array)
         if model is None:
-            model = RandomForestWithInstances(types=types, bounds=bounds,
+            model = RandomForestWithInstances(types=types, 
+                                              bounds=bounds,
                                               instance_features=scenario.feature_array,
                                               seed=rng.randint(MAXINT),
-                                              pca_components=scenario.PCA_DIM)
+                                              pca_components=scenario.PCA_DIM,
+                                              unlog_y=scenario.run_obj == "runtime")
         # initial acquisition function
         if acquisition_function is None:
-            if scenario.run_obj == "runtime":
-                acquisition_function = LogEI(model=model)
-            else:
-                acquisition_function = EI(model=model)
+            acquisition_function = EI(model=model)
+
         # inject model if necessary
         if acquisition_function.model is None:
             acquisition_function.model = model
@@ -182,7 +203,7 @@ class SMAC(object):
         # initialize optimizer on acquisition function
         if acquisition_function_optimizer is None:
             acquisition_function_optimizer = InterleavedLocalAndRandomSearch(
-                acquisition_function, scenario.cs
+                acquisition_function, scenario.cs, np.random.RandomState(seed=rng.randint(MAXINT))
             )
         elif not isinstance(
                 acquisition_function_optimizer,
@@ -366,8 +387,7 @@ class SMAC(object):
         else:
             self.solver = smbo_class(**smbo_args)
 
-    @staticmethod
-    def _get_rng(rng):
+    def _get_rng(self, rng):
         """Initialize random number generator
 
         If rng is None, initialize a new generator
@@ -384,13 +404,14 @@ class SMAC(object):
         """
         # initialize random number generator
         if rng is None:
-            num_run = np.random.randint(1234567980)
+            self.logger.debug('no rng given: using default seed of 1')
+            num_run = 1
             rng = np.random.RandomState(seed=num_run)
         elif isinstance(rng, int):
             num_run = rng
             rng = np.random.RandomState(seed=rng)
         elif isinstance(rng, np.random.RandomState):
-            num_run = rng.randint(1234567980)
+            num_run = rng.randint(MAXINT)
             rng = rng
         else:
             raise TypeError('Unknown type %s for argument rng. Only accepts '
@@ -448,9 +469,9 @@ class SMAC(object):
         config_mode: str or list<Configuration>
             string or directly a list of Configuration
             str from [def, inc, def+inc, wallclock_time, cpu_time, all]
-                time evaluates at cpu- or wallclock-timesteps of:
-                [max_time/2^0, max_time/2^1, max_time/2^3, ..., default]
-                with max_time being the highest recorded time
+            time evaluates at cpu- or wallclock-timesteps of:
+            [max_time/2^0, max_time/2^1, max_time/2^3, ..., default]
+            with max_time being the highest recorded time
         instance_mode: string
             what instances to use for validation, from [train, test, train+test]
         repetitions: int