+++ test/test_cli/test_restore_state.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import unittest
+from nose.plugins.attrib import attr
 import shutil
 
 import numpy as np
@@ -26,17 +27,23 @@ class TestSMACCLI(unittest.TestCase):
         self.current_dir = os.getcwd()
         os.chdir(base_directory)
 
-        self.output_one = "test/test_files/test_restore_state/run_1"  # From scenario_one.txt
-        self.output_two = "test/test_files/test_restored_state/run_1" # From scenario_two.txt
+        output_one_dir = "test/test_files/test_restore_state"  # From scenario_one.txt
+        self.output_one = output_one_dir + "/run_1"
+        output_two_dir = "test/test_files/test_restored_state"  # From scenario_two.txt
+        self.output_two = output_two_dir + "/run_1"
         self.smaccli = SMACCLI()
         self.scenario_one = "test/test_files/restore_scenario_one.txt"
         self.scenario_two = "test/test_files/restore_scenario_two.txt"
+        self.output_dirs = [output_one_dir, output_two_dir]
 
     def tearDown(self):
-        shutil.rmtree(self.output_one, ignore_errors=True)
-        shutil.rmtree(self.output_two, ignore_errors=True)
+        for output_dir in self.output_dirs:
+            if output_dir:
+                shutil.rmtree(output_dir, ignore_errors=True)
+                #pass
         os.chdir(self.current_dir)
 
+    @attr('slow')
     def test_run_and_restore(self):
         """
         Testing basic restore functionality.
@@ -74,6 +81,7 @@ class TestSMACCLI(unittest.TestCase):
         # Recorded runs but no incumbent.
         stats.ta_runs = 10
         smac = SMAC(scen, stats=stats, rng=np.random.RandomState(42))
+        self.output_dirs.append(scen.output_dir)
         self.assertRaises(ValueError, smac.optimize)
         # Incumbent but no recoreded runs.
         incumbent = cs.get_default_configuration()
@@ -93,7 +101,12 @@ class TestSMACCLI(unittest.TestCase):
         # Increase limit and run for 10 (so 5 more) by using restore_state
         testargs = ["python", "scripts/smac", "--restore_state",
                     self.output_one, "--scenario_file",
-                    self.scenario_one, "--verbose", "DEBUG"]
+                    self.scenario_two, "--verbose", "DEBUG"]
         with mock.patch.object(sys, 'argv', testargs):
             self.smaccli.main_cli()
+        self.assertTrue(os.path.exists(self.output_one))
+        self.assertFalse(os.path.exists(self.output_one + '.OLD'))
+        self.assertTrue(os.path.exists(self.output_two))
+        self.assertFalse(os.path.exists(self.output_two + '.OLD'))
+